--- src/app/(route)/todo/apis/mock_todo.ts
@@ -1,16 +1,0 @@
-export const mockTodoResponse = {
-  todos: [
-    {
-      status: false,
-      createdAt: '2024-01-24T00:00:00.000+09:00',
-      text: '바이올린하기',
-      id: '95cfc70f-69f5-4a88-8960-dda089dae24a',
-    },
-    {
-      status: false,
-      createdAt: '2024-01-24T00:00:00.000+09:00',
-      text: '물 사기',
-      id: 'e3489f61-59a3-490d-9adc-403d595feaa2',
-    },
-  ],
-};
